feat(wealth-library): phase 0 — extract MacroReviewBody standalone
Extracts the weekly Macro Committee Review body (regional score
deltas, regime panels, staleness alerts, decision rationale) into
lib/components/library/readers/MacroReviewBody.svelte. Strict
prop contract: only `reviewId`. Region highlight is read from the
URL search params at runtime so the body works inside any host
shell without depending on `$page`.

Hardcoded badge palette literals (#f3f4f6, #dcfce7, #dbeafe,
#fee2e2, #166534, #1e40af, #991b1b) replaced with the
Screener dark-mode token palette via color-mix() against
--ii-success / --ii-brand-primary / --ii-danger so the body
slots into any institutional surface.

The legacy route at market/reviews/[reviewId] becomes a thin
shell: a back link plus <MacroReviewBody {reviewId} />. The
+page.server.ts loader is unchanged so deep-linking still works.
